Update.
None.

+++ TC264Project/SmartCar/LQ_TC264DA_LIB_ADS/App/Usr/parameter.h
@@ -49,7 +49,7 @@ typedef struct
 
       uint AI_State;
 
-      uint16_t CarState;
+      uint CarState;
 
       const uint CarMode;
 }data_t;
